% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedocumentr-package.R
\docType{package}
\name{filedocumentr-package}
\alias{filedocumentr}
\alias{filedocumentr-package}
\title{filedocumentr: File documentation for data directories}
\description{
This package contains README template file to be used to document metadate associated with files in a data directory.
}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID})

}
\keyword{internal}
