% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-entries.R
\name{library_entry}
\alias{library_entry}
\title{Lower level helper function to enable saving of intermediate data entries.}
\usage{
library_entry(
  .library_name,
  ...,
  .create_dir = TRUE,
  .alt_path = NULL,
  .save = TRUE,
  .primary_key = TRUE
)
}
\arguments{
\item{.library_name}{A character string for the name of the table.}

\item{...}{The variables entered in the the table.}

\item{.create_dir}{Boolean indicating whether a separated directory is
generated.}

\item{.alt_path}{Alternative path to the directory.}

\item{.save}{Boolean, whether to save the entry as an .RDS file.}

\item{.primary_key}{Equivalent to a serial primary key in SQL for uniqueness
of observations.}
}
\value{
A \code{tibble:\link[tibble:tibble]{tibble()}} alongside a .RDS file
with the entries.
}
\description{
Lower level helper function to enable saving of intermediate data entries.
}
