% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{use_filedocumentr}
\alias{use_filedocumentr}
\title{Document files structure in data directory}
\usage{
use_filedocumentr(name = "README")
}
\value{
Rmd file with template for common metadata associated with  data files.
}
\description{
Create an overview file of the data consisted within the directory by
documenting some minimal information, such as author, date of data generation
, instrumentation and laboratory.
}
\examples{
# generate README in current directory
use_filedocumentr()
}
